+++ internal/pkg/warewulfd/efi.go
@@ -44,6 +44,14 @@ func HandleEfiBoot(w http.ResponseWriter, req *http.Request) {
 			return
 		}
 	case "grub.cfg":
+		// this is the first stage fot the node in the boot process where the
+		// host is involved, so we reset the SentLog and *prepend* is with the
+		// grub.cfg which was sent out via tftp, so clear the logs and add as
+		// first entry the grub.cfg we sent out via tftp
+		ctx.tpm.ClearLogs()
+		if err = ctx.tpm.Update(path.Join(ctx.conf.TFTP.TftpRoot, "warewulf/grub.cfg"), ""); err != nil {
+			wwlog.Warn("couldn't update TPM log with grub.cfg sent by tftp", err)
+		}
 		stageFile = path.Join(ctx.conf.Paths.Sysconfdir, "warewulf/grub/grub.cfg.ww")
 		tmplData = buildTemplateVars(ctx.conf, ctx.rinfo, ctx.remoteNode)
 		if !util.IsFile(stageFile) {
